requirements.txt
boto==2.48.0
boto3==1.6.6
botocore==1.9.6
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.7
django-appconf==1.0.2
django-background-image==0.0.1
django-bootstrap3==10.0.1
django-bootstrap3-datetimepicker-3==2.6.0
django-heroku==0.3.1
django-storages==1.6.6
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.3
Pillow==5.2.0
psycopg2==2.7.5
python-dateutil==2.6.1
pytz==2018.5
rcssmin==1.0.6
rjsmin==1.0.12
s3transfer==0.1.13
six==1.11.0
virtualenv==16.0.0
whitenoise==3.3.1
